requirements.txt
gradio
yfinance
plotly
pandas
numpy
python-dotenv
google-generativeai
scikit-learn
scipy        